un type si niste importuri
