+++ MarketMinds.Test/Services/AuctionProductsServiceTest/AuctionProductsServiceTest.cs
@@ -188,67 +188,5 @@ namespace MarketMinds.Tests.Services.AuctionProductsServiceTest
             Assert.That(auctionProductsRepositoryMock.GetProducts().Count, Is.EqualTo(1));
             Assert.That(auctionProductsRepositoryMock.GetProducts()[0], Is.EqualTo(testAuction));
         }
-
-        [Test]
-        public void TestGetTimeLeft_AuctionActive_ReturnsFormattedTimeLeft()
-        {
-            // Set the auction end date to a specific time in the future
-            testAuction.EndAuctionDate = DateTime.Now.AddDays(2).AddHours(3).AddMinutes(45).AddSeconds(30);
-
-            string timeLeft = auctionProductsService.GetTimeLeft(testAuction);
-
-            // Verify the format is correct (dd:hh:mm:ss)
-            Assert.That(timeLeft, Does.Match(@"^\d{2}:\d{2}:\d{2}:\d{2}$"));
-
-            // Verify the time left starts with "02:" for 2 days
-            Assert.That(timeLeft, Does.StartWith("02:"));
-        }
-
-        [Test]
-        public void TestGetTimeLeft_AuctionEnded_ReturnsAuctionEndedMessage()
-        {
-            // Set the auction end date to a time in the past
-            testAuction.EndAuctionDate = DateTime.Now.AddHours(-5);
-
-            string timeLeft = auctionProductsService.GetTimeLeft(testAuction);
-
-            Assert.That(timeLeft, Is.EqualTo("Auction Ended"));
-        }
-
-        [Test]
-        public void TestIsAuctionEnded_EndDateInFuture_ReturnsFalse()
-        {
-            // Set the auction end date to a time in the future
-            testAuction.EndAuctionDate = DateTime.Now.AddHours(1);
-
-            bool isEnded = auctionProductsService.IsAuctionEnded(testAuction);
-
-            Assert.That(isEnded, Is.False);
-        }
-
-        [Test]
-        public void TestIsAuctionEnded_EndDateInPast_ReturnsTrue()
-        {
-            // Set the auction end date to a time in the past
-            testAuction.EndAuctionDate = DateTime.Now.AddMinutes(-30);
-
-            bool isEnded = auctionProductsService.IsAuctionEnded(testAuction);
-
-            Assert.That(isEnded, Is.True);
-        }
-
-        [Test]
-        public void TestIsAuctionEnded_EndDateIsNow_ReturnsTrue()
-        {
-            // Set the auction end date to current time (or as close as possible)
-            testAuction.EndAuctionDate = DateTime.Now;
-
-            bool isEnded = auctionProductsService.IsAuctionEnded(testAuction);
-
-            // Since DateTime.Now may change between setting and checking,
-            // and because of precision issues, we need to be careful with this assertion
-            // In reality, it should return true since DateTime.Now >= DateTime.Now should be true
-            Assert.That(isEnded, Is.True);
-        }
     }
 }
