+++ MarketMinds.Test/Services/BasketServiceTest/BasketServiceTest.cs
@@ -534,243 +534,5 @@ namespace MarketMinds.Tests.Services.BasketServiceTest
             // Attempt to calculate totals with invalid basket ID
             Assert.Throws<ArgumentException>(() => basketService.CalculateBasketTotals(invalidBasketId, null));
         }
-
-        [Test]
-        public void TestValidateQuantityInput_ValidNumber_ReturnsTrue()
-        {
-            bool result = basketService.ValidateQuantityInput("5", out int quantity);
-
-            Assert.That(result, Is.True);
-            Assert.That(quantity, Is.EqualTo(5));
-        }
-
-        [Test]
-        public void TestValidateQuantityInput_Zero_ReturnsTrue()
-        {
-            bool result = basketService.ValidateQuantityInput("0", out int quantity);
-
-            Assert.That(result, Is.True);
-            Assert.That(quantity, Is.EqualTo(0));
-        }
-
-        [Test]
-        public void TestValidateQuantityInput_NegativeNumber_ReturnsFalse()
-        {
-            bool result = basketService.ValidateQuantityInput("-5", out int quantity);
-
-            Assert.That(result, Is.False);
-        }
-
-        [Test]
-        public void TestValidateQuantityInput_NonNumeric_ReturnsFalse()
-        {
-            bool result = basketService.ValidateQuantityInput("abc", out int quantity);
-
-            Assert.That(result, Is.False);
-        }
-
-        [Test]
-        public void TestValidateQuantityInput_EmptyString_ReturnsFalse()
-        {
-            bool result = basketService.ValidateQuantityInput("", out int quantity);
-
-            Assert.That(result, Is.False);
-        }
-
-        [Test]
-        public void TestGetLimitedQuantity_BelowMax_ReturnsSameValue()
-        {
-            int quantity = BasketService.MaxQuantityPerItem - 1;
-
-            int result = basketService.GetLimitedQuantity(quantity);
-
-            Assert.That(result, Is.EqualTo(quantity));
-        }
-
-        [Test]
-        public void TestGetLimitedQuantity_AboveMax_ReturnsMax()
-        {
-            int quantity = BasketService.MaxQuantityPerItem + 10;
-
-            int result = basketService.GetLimitedQuantity(quantity);
-
-            Assert.That(result, Is.EqualTo(BasketService.MaxQuantityPerItem));
-        }
-
-        [Test]
-        public void TestGetLimitedQuantity_EqualToMax_ReturnsMax()
-        {
-            int quantity = BasketService.MaxQuantityPerItem;
-
-            int result = basketService.GetLimitedQuantity(quantity);
-
-            Assert.That(result, Is.EqualTo(BasketService.MaxQuantityPerItem));
-        }
-
-        [Test]
-        public void TestDecreaseProductQuantity_QuantityGreaterThanOne_DecreasesQuantity()
-        {
-            // Create a basket and add an item to it
-            var basket = basketRepositoryMock.GetBasketByUserId(testUserId);
-            basketRepositoryMock.AddItemToBasket(basket.Id, testProductId, 3);
-
-            basketService.DecreaseProductQuantity(testUserId, testProductId);
-
-            Assert.That(basketRepositoryMock.GetUpdateItemCount(), Is.EqualTo(1));
-        }
-
-        [Test]
-        public void TestDecreaseProductQuantity_QuantityOne_RemovesItem()
-        {
-            // Create a basket and add an item to it
-            var basket = basketRepositoryMock.GetBasketByUserId(testUserId);
-            basketRepositoryMock.AddItemToBasket(basket.Id, testProductId, 1);
-
-            basketService.DecreaseProductQuantity(testUserId, testProductId);
-
-            Assert.That(basketRepositoryMock.GetRemoveItemCount(), Is.EqualTo(1));
-        }
-
-        [Test]
-        public void TestDecreaseProductQuantity_InvalidUserId_ThrowsException()
-        {
-            Exception ex = null;
-            try
-            {
-                basketService.DecreaseProductQuantity(0, testProductId);
-            }
-            catch (Exception e)
-            {
-                ex = e;
-            }
-
-            Assert.That(ex, Is.Not.Null);
-            Assert.That(ex, Is.TypeOf<ArgumentException>());
-            Assert.That(ex.Message, Is.EqualTo("Invalid user ID"));
-        }
-
-        [Test]
-        public void TestDecreaseProductQuantity_InvalidProductId_ThrowsException()
-        {
-            Exception ex = null;
-            try
-            {
-                basketService.DecreaseProductQuantity(testUserId, 0);
-            }
-            catch (Exception e)
-            {
-                ex = e;
-            }
-
-            Assert.That(ex, Is.Not.Null);
-            Assert.That(ex, Is.TypeOf<ArgumentException>());
-            Assert.That(ex.Message, Is.EqualTo("Invalid product ID"));
-        }
-
-        [Test]
-        public void TestDecreaseProductQuantity_ItemNotFound_ThrowsException()
-        {
-            // Create a basket but add a different product
-            var basket = basketRepositoryMock.GetBasketByUserId(testUserId);
-            basketRepositoryMock.AddItemToBasket(basket.Id, 999, 1);
-
-            Exception ex = null;
-            try
-            {
-                basketService.DecreaseProductQuantity(testUserId, testProductId);
-            }
-            catch (Exception e)
-            {
-                ex = e;
-            }
-
-            Assert.That(ex, Is.Not.Null);
-            Assert.That(ex, Is.TypeOf<InvalidOperationException>());
-            // Fix: match the actual wrapper exception message
-            Assert.That(ex.Message, Is.EqualTo("Could not decrease quantity: Item not found in basket"));
-        }
-
-        [Test]
-        public void TestIncreaseProductQuantity_BelowMax_IncreasesQuantity()
-        {
-            // Create a basket and add an item to it
-            var basket = basketRepositoryMock.GetBasketByUserId(testUserId);
-            basketRepositoryMock.AddItemToBasket(basket.Id, testProductId, BasketService.MaxQuantityPerItem - 1);
-
-            basketService.IncreaseProductQuantity(testUserId, testProductId);
-
-            Assert.That(basketRepositoryMock.GetUpdateItemCount(), Is.EqualTo(1));
-        }
-
-        [Test]
-        public void TestIncreaseProductQuantity_AtMax_DoesNotExceedMax()
-        {
-            // Create a basket and add an item to it
-            var basket = basketRepositoryMock.GetBasketByUserId(testUserId);
-            basketRepositoryMock.AddItemToBasket(basket.Id, testProductId, BasketService.MaxQuantityPerItem);
-
-            basketService.IncreaseProductQuantity(testUserId, testProductId);
-
-            Assert.That(basketRepositoryMock.GetUpdateItemCount(), Is.EqualTo(1));
-        }
-
-        [Test]
-        public void TestIncreaseProductQuantity_InvalidUserId_ThrowsException()
-        {
-            Exception ex = null;
-            try
-            {
-                basketService.IncreaseProductQuantity(0, testProductId);
-            }
-            catch (Exception e)
-            {
-                ex = e;
-            }
-
-            Assert.That(ex, Is.Not.Null);
-            Assert.That(ex, Is.TypeOf<ArgumentException>());
-            Assert.That(ex.Message, Is.EqualTo("Invalid user ID"));
-        }
-
-        [Test]
-        public void TestIncreaseProductQuantity_InvalidProductId_ThrowsException()
-        {
-            Exception ex = null;
-            try
-            {
-                basketService.IncreaseProductQuantity(testUserId, 0);
-            }
-            catch (Exception e)
-            {
-                ex = e;
-            }
-
-            Assert.That(ex, Is.Not.Null);
-            Assert.That(ex, Is.TypeOf<ArgumentException>());
-            Assert.That(ex.Message, Is.EqualTo("Invalid product ID"));
-        }
-
-        [Test]
-        public void TestIncreaseProductQuantity_ItemNotFound_ThrowsException()
-        {
-            // Create a basket but add a different product
-            var basket = basketRepositoryMock.GetBasketByUserId(testUserId);
-            basketRepositoryMock.AddItemToBasket(basket.Id, 999, 1);
-
-            Exception ex = null;
-            try
-            {
-                basketService.IncreaseProductQuantity(testUserId, testProductId);
-            }
-            catch (Exception e)
-            {
-                ex = e;
-            }
-
-            Assert.That(ex, Is.Not.Null);
-            Assert.That(ex, Is.TypeOf<InvalidOperationException>());
-            // Fix: match the actual wrapper exception message
-            Assert.That(ex.Message, Is.EqualTo("Could not increase quantity: Item not found in basket"));
-        }
     }
 }
