+++ MarketMinds.Test/Services/ProductService/ProductServiceTest.cs
@@ -335,258 +335,7 @@ namespace MarketMinds.Test.Services.ProductTagService
             Assert.That(result.Count, Is.EqualTo(3));
         }
 
-        [Test]
-        public void GetSortedFilteredProducts_WithSpecificConditionFilter_ShouldReturnMatchingProducts()
-        {
-            // Arrange
-            var mockRepository = new ProductRepositoryMock();
-            var productService = new MarketMinds.Services.ProductTagService.ProductService(mockRepository);
 
-            var newCondition = new ProductCondition(1, "New", "Brand new");
-            var usedCondition = new ProductCondition(2, "Used", "Used item");
-            var refurbishedCondition = new ProductCondition(3, "Refurbished", "Refurbished item");
-
-            var product1 = CreateSampleProduct(1, "Laptop", condition: newCondition);
-            var product2 = CreateSampleProduct(2, "Phone", condition: usedCondition);
-            var product3 = CreateSampleProduct(3, "Tablet", condition: refurbishedCondition);
-            var product4 = CreateSampleProduct(4, "Desktop", condition: newCondition);
-
-            mockRepository.AddProduct(product1);
-            mockRepository.AddProduct(product2);
-            mockRepository.AddProduct(product3);
-            mockRepository.AddProduct(product4);
-
-            var selectedConditions = new List<ProductCondition> { newCondition };
-
-            // Act
-            var result = productService.GetSortedFilteredProducts(selectedConditions, null, null, null, null);
-
-            // Assert
-            Assert.That(result.Count, Is.EqualTo(2));
-            Assert.That(result.All(p => p.Condition.Id == newCondition.Id), Is.True);
-            Assert.That(result.Select(p => p.Title).Contains("Laptop"), Is.True);
-            Assert.That(result.Select(p => p.Title).Contains("Desktop"), Is.True);
-        }
-
-        [Test]
-        public void GetSortedFilteredProducts_WithSpecificCategoryFilter_ShouldReturnMatchingProducts()
-        {
-            // Arrange
-            var mockRepository = new ProductRepositoryMock();
-            var productService = new MarketMinds.Services.ProductTagService.ProductService(mockRepository);
-
-            var electronicsCategory = new ProductCategory(1, "Electronics", "Electronic devices");
-            var clothingCategory = new ProductCategory(2, "Clothing", "Apparel items");
-
-            var product1 = CreateSampleProduct(1, "Laptop", category: electronicsCategory);
-            var product2 = CreateSampleProduct(2, "T-shirt", category: clothingCategory);
-            var product3 = CreateSampleProduct(3, "Jeans", category: clothingCategory);
-            var product4 = CreateSampleProduct(4, "Headphones", category: electronicsCategory);
-
-            mockRepository.AddProduct(product1);
-            mockRepository.AddProduct(product2);
-            mockRepository.AddProduct(product3);
-            mockRepository.AddProduct(product4);
-
-            var selectedCategories = new List<ProductCategory> { clothingCategory };
-
-            // Act
-            var result = productService.GetSortedFilteredProducts(null, selectedCategories, null, null, null);
-
-            // Assert
-            Assert.That(result.Count, Is.EqualTo(2));
-            Assert.That(result.All(p => p.Category.Id == clothingCategory.Id), Is.True);
-            Assert.That(result.Select(p => p.Title).Contains("T-shirt"), Is.True);
-            Assert.That(result.Select(p => p.Title).Contains("Jeans"), Is.True);
-        }
-
-        [Test]
-        public void GetSortedFilteredProducts_WithSpecificTagFilter_ShouldReturnMatchingProducts()
-        {
-            // Arrange
-            var mockRepository = new ProductRepositoryMock();
-            var productService = new MarketMinds.Services.ProductTagService.ProductService(mockRepository);
-
-            var wirelessTag = new ProductTag(1, "Wireless");
-            var bluetoothTag = new ProductTag(2, "Bluetooth");
-            var waterproofTag = new ProductTag(3, "Waterproof");
-
-            var product1 = CreateSampleProduct(1, "Headphones");
-            product1.Tags.Add(wirelessTag);
-            product1.Tags.Add(bluetoothTag);
-
-            var product2 = CreateSampleProduct(2, "Speaker");
-            product2.Tags.Add(bluetoothTag);
-            product2.Tags.Add(waterproofTag);
-
-            var product3 = CreateSampleProduct(3, "Mouse");
-            product3.Tags.Add(wirelessTag);
-
-            var product4 = CreateSampleProduct(4, "Keyboard");
-            product4.Tags.Add(wirelessTag);
-
-            mockRepository.AddProduct(product1);
-            mockRepository.AddProduct(product2);
-            mockRepository.AddProduct(product3);
-            mockRepository.AddProduct(product4);
-
-            var selectedTags = new List<ProductTag> { wirelessTag };
-
-            // Act
-            var result = productService.GetSortedFilteredProducts(null, null, selectedTags, null, null);
-
-            // Assert
-            Assert.That(result.Count, Is.EqualTo(3));
-            Assert.That(result.All(p => p.Tags.Any(t => t.Id == wirelessTag.Id)), Is.True);
-            Assert.That(result.Select(p => p.Title).Contains("Headphones"), Is.True);
-            Assert.That(result.Select(p => p.Title).Contains("Mouse"), Is.True);
-            Assert.That(result.Select(p => p.Title).Contains("Keyboard"), Is.True);
-        }
-
-        [Test]
-        public void GetSortedFilteredProducts_WithSearchQuery_ShouldReturnMatchingProducts()
-        {
-            // Arrange
-            var mockRepository = new ProductRepositoryMock();
-            var productService = new MarketMinds.Services.ProductTagService.ProductService(mockRepository);
-
-            mockRepository.AddProduct(CreateSampleProduct(1, "Gaming Laptop"));
-            mockRepository.AddProduct(CreateSampleProduct(2, "Business Laptop"));
-            mockRepository.AddProduct(CreateSampleProduct(3, "Gaming Desktop"));
-            mockRepository.AddProduct(CreateSampleProduct(4, "Smartphone"));
-
-            // Act
-            var result = productService.GetSortedFilteredProducts(null, null, null, null, "laptop");
-
-            // Assert
-            Assert.That(result.Count, Is.EqualTo(2));
-            Assert.That(result.All(p => p.Title.ToLower().Contains("laptop")), Is.True);
-            Assert.That(result.Select(p => p.Title).Contains("Gaming Laptop"), Is.True);
-            Assert.That(result.Select(p => p.Title).Contains("Business Laptop"), Is.True);
-        }
-
-        [Test]
-        public void GetSortedFilteredProducts_WithTitleSorting_ShouldSortByTitle()
-        {
-            // Arrange
-            var mockRepository = new ProductRepositoryMock();
-            var productService = new MarketMinds.Services.ProductTagService.ProductService(mockRepository);
-
-            mockRepository.AddProduct(CreateSampleProduct(1, "Zebra"));
-            mockRepository.AddProduct(CreateSampleProduct(2, "Apple"));
-            mockRepository.AddProduct(CreateSampleProduct(3, "Banana"));
-            mockRepository.AddProduct(CreateSampleProduct(4, "Xylophone"));
-
-            var sortCondition = new ProductSortType("Title", "Title", true);
-
-            // Act
-            var result = productService.GetSortedFilteredProducts(null, null, null, sortCondition, null);
-
-            // Assert
-            Assert.That(result.Count, Is.EqualTo(4));
-            Assert.That(result[0].Title, Is.EqualTo("Apple"));
-            Assert.That(result[1].Title, Is.EqualTo("Banana"));
-            Assert.That(result[2].Title, Is.EqualTo("Xylophone"));
-            Assert.That(result[3].Title, Is.EqualTo("Zebra"));
-        }
-
-        [Test]
-        public void GetSortedFilteredProducts_WithTitleSortingDescending_ShouldSortByTitleDescending()
-        {
-            // Arrange
-            var mockRepository = new ProductRepositoryMock();
-            var productService = new MarketMinds.Services.ProductTagService.ProductService(mockRepository);
-
-            mockRepository.AddProduct(CreateSampleProduct(1, "Zebra"));
-            mockRepository.AddProduct(CreateSampleProduct(2, "Apple"));
-            mockRepository.AddProduct(CreateSampleProduct(3, "Banana"));
-            mockRepository.AddProduct(CreateSampleProduct(4, "Xylophone"));
-
-            var sortCondition = new ProductSortType("Title", "Title", false);
-
-            // Act
-            var result = productService.GetSortedFilteredProducts(null, null, null, sortCondition, null);
-
-            // Assert
-            Assert.That(result.Count, Is.EqualTo(4));
-            Assert.That(result[0].Title, Is.EqualTo("Zebra"));
-            Assert.That(result[1].Title, Is.EqualTo("Xylophone"));
-            Assert.That(result[2].Title, Is.EqualTo("Banana"));
-            Assert.That(result[3].Title, Is.EqualTo("Apple"));
-        }
-
-        [Test]
-        public void GetSortedFilteredProducts_WithMultipleFilters_ShouldApplyAllFilters()
-        {
-            // Arrange
-            var mockRepository = new ProductRepositoryMock();
-            var productService = new MarketMinds.Services.ProductTagService.ProductService(mockRepository);
-
-            var electronicsCategory = new ProductCategory(1, "Electronics", "Electronic devices");
-            var newCondition = new ProductCondition(1, "New", "Brand new");
-            var wirelessTag = new ProductTag(1, "Wireless");
-
-            var product1 = CreateSampleProduct(1, "Wireless Gaming Mouse", electronicsCategory, newCondition);
-            product1.Tags.Add(wirelessTag);
-
-            var product2 = CreateSampleProduct(2, "Wired Gaming Mouse", electronicsCategory, newCondition);
-
-            var product3 = CreateSampleProduct(3, "Wireless Keyboard", electronicsCategory, newCondition);
-            product3.Tags.Add(wirelessTag);
-
-            mockRepository.AddProduct(product1);
-            mockRepository.AddProduct(product2);
-            mockRepository.AddProduct(product3);
-
-            var selectedCategories = new List<ProductCategory> { electronicsCategory };
-            var selectedConditions = new List<ProductCondition> { newCondition };
-            var selectedTags = new List<ProductTag> { wirelessTag };
-
-            // Act
-            var result = productService.GetSortedFilteredProducts(selectedConditions, selectedCategories, selectedTags, null, "mouse");
-
-            // Assert
-            Assert.That(result.Count, Is.EqualTo(1));
-            Assert.That(result[0].Title, Is.EqualTo("Wireless Gaming Mouse"));
-        }
-
-        [Test]
-        public void GetSortedFilteredProducts_WithEmptyResults_ShouldReturnEmptyList()
-        {
-            // Arrange
-            var mockRepository = new ProductRepositoryMock();
-            var productService = new MarketMinds.Services.ProductTagService.ProductService(mockRepository);
-
-            var product1 = CreateSampleProduct(1, "Laptop");
-            var product2 = CreateSampleProduct(2, "Desktop");
-
-            mockRepository.AddProduct(product1);
-            mockRepository.AddProduct(product2);
-
-            // Act
-            var result = productService.GetSortedFilteredProducts(null, null, null, null, "nonexistent");
-
-            // Assert
-            Assert.That(result.Count, Is.EqualTo(0));
-        }
-
-        [Test]
-        public void GetSortedFilteredProducts_WithNullRepository_ShouldHandleGracefully()
-        {
-            // This test directly calls the static method without a repository
-
-            // Arrange
-            var product1 = CreateSampleProduct(1, "Product A");
-            var product2 = CreateSampleProduct(2, "Product B");
-            var products = new List<Product> { product1, product2 };
-
-            // Act
-            var result = MarketMinds.Services.ProductTagService.ProductService.GetSortedFilteredProducts(
-                products, null, null, null, null, null);
-
-            // Assert
-            Assert.That(result.Count, Is.EqualTo(2));
-        }
 
         private Product CreateSampleProduct(int id, string title, ProductCategory category = null, ProductCondition condition = null)
         {
